third_party/WebKit/LayoutTests/fast/dom/shadow/shadow-aware-shadow-root-expected.txt

Tests for ShadowAware.shadowRoot property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS host.shadowRoot is null
PASS root1 is host.shadowRoot
PASS root2 is host.shadowRoot
PASS hostWithUAShadow.shadowRoot is null
PASS successfullyParsed is true

TEST COMPLETE
